+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed keep_res regular expression pattern from options
 - Instead of caching if a DRBD .res file exists, actually check if the file exists
 
+### Fixed
+
+- Attempt to fix race-condition "Received unknown storage resource from satellite"
+
 ## [1.32.2] - 2025-10-01
 
 ### Changed